+++ utils.js
@@ -581,3 +581,29 @@ export async function fetch_teams_bet(sportname, username) {
     }).catch(err => console.log(err));
     return fetch_teams;
 }
+
+export function lock_unlock(lock, setLock, sportname) {
+    const controller = new AbortController();
+    let type = "";
+    if (lock) {
+
+        type = "unlock";
+    }
+    else {
+
+        type = "lock";
+    }
+    const timeoutId = setTimeout(() => controller.abort(), 10000);
+    fetch("https://applijo.freeddns.org/locksport", { signal: controller.signal, method: "POST", body: JSON.stringify({ "version": version, "sport": sportname, "type": type }) }).then(r => {
+        if (r.status == 200) {
+
+            setLock(!lock);
+        }
+        else {
+            let msg = "Server reply :" + r.status
+            alert(msg);
+        }
+    }
+    ).catch(err => console.error(err))
+
+}
